Fix field tags
