+++ client/src/styles/colors.scss
@@ -14,3 +14,5 @@ $recentlyActive: #e9f4e7;
 $active: #e7f2fc;
 $inactive: #fef5e7;
 $pending: #fef0db;
+$approved: #e9f4e7;
+$declined: #ffe9e7;
