+++ client/src/styles/styles.scss
@@ -73,6 +73,16 @@ body {
   background-color: $pending !important;
 }
 
+.approved-tag-color,
+.approved-review-tag-color {
+  background-color: $approved !important;
+}
+
+.declined-tag-color,
+.declined-review-tag-color {
+  background-color: $declined !important;
+}
+
 .default-btn {
   background-color: $secondaryColor !important;
   color: white !important;
